Report error if there is no required data

+++ lib/tiny-classifier/command/generate-classifier.rb
@@ -34,6 +34,9 @@ module TinyClassifier
   
       def run
         super
+        raise NoTrainingData.new unless data_file_path.exist?
+        raise InvalidOutputDir.new unless prepare_output_file_path.parent.exist?
+
         FileUtils.mkdir_p(output_file_path.parent)
         File.open(output_file_path, "w") do |file|
           file.puts("#!/usr/bin/env ruby")
